+++ IP (Introdução a Programação)/Lista - 3/Exercicio - 3.c
@@ -3,15 +3,15 @@
 
 /* run this program using the console pauser or add your own getch, system("pause") or input loop */
 
-void verificarRepeticoesNumero(int numero, int digitoProcurado) {
-    int repeticoes = 0;
+void verificarRepeticoesNumero(short int numero, short int digitoProcurado) {
+    short int repeticoes = 0;
 
     while (numero >= 10) {
-        int ultimoDigitoNumero = numero % 10;
+        short int ultimoDigitoNumero = numero % 10;
         
         numero = numero / 10;
 
-        int penultimoDigitoNumero = numero % 10;
+        short int penultimoDigitoNumero = numero % 10;
 
         if (ultimoDigitoNumero == penultimoDigitoNumero && ultimoDigitoNumero == digitoProcurado) {
             repeticoes++;
@@ -19,7 +19,7 @@ void verificarRepeticoesNumero(int numero, int digitoProcurado) {
     }
 
     if (repeticoes != 0) {
-        printf("Numero de repeticoes: %d\n", repeticoes);
+        printf("Numero de repeticoes: %hd\n", repeticoes);
     } else {
         printf("Nao ha repeticoes");
     }
@@ -27,15 +27,15 @@ void verificarRepeticoesNumero(int numero, int digitoProcurado) {
 
 int main(int argc, char const *argv[])
 {
-    int numero, digito;
+    short int numero, digito;
 
     do {
-    printf("Digite o numero inteiro positivo para verificar repeticoes nele: ");
-    scanf("%d", &numero);
+        printf("Digite o numero inteiro positivo para verificar repeticoes nele: ");
+        scanf("%hd", &numero);
     } while (numero <= 0);
 
     printf("Digite o digito de um unico digito: ");
-    scanf("%d", &digito);
+    scanf("%hd", &digito);
 
     verificarRepeticoesNumero(numero, digito);
